Added --dry-run mode, ability to test render without clobbering.
--HG--
extra : source : 042b7615db86c3456eafe00caaa86085954ec5d9

+++ poole.py
@@ -515,13 +515,14 @@ def build(project, opts):
             sys.exit(1)
 
     # prepare output directory
-    for fod in glob.glob(opj(dir_out, "*")):
-        if os.path.isdir(fod):
-            shutil.rmtree(fod)
-        else:
-            os.remove(fod)
-    if not opx(dir_out):
-        os.mkdir(dir_out)
+    if not opts.dry_run:
+        for fod in glob.glob(opj(dir_out, "*")):
+            if os.path.isdir(fod):
+                shutil.rmtree(fod)
+            else:
+                os.remove(fod)
+        if not opx(dir_out):
+            os.mkdir(dir_out)
 
     # macro module
     fname = opj(opts.project, "macros.py")
@@ -550,11 +551,12 @@ def build(project, opts):
 
     for cwd, dirs, files in os.walk(dir_in.decode(opts.filename_enc)):
         cwd_site = cwd[len(dir_in):].lstrip(os.path.sep)
-        for sdir in dirs[:]:
-            if re.search(opts.ignore, opj(cwd_site, sdir)):
-                dirs.remove(sdir)
-            else:
-                os.mkdir(opj(dir_out, cwd_site, sdir))
+        if not opts.dry_run:
+            for sdir in dirs[:]:
+                if re.search(opts.ignore, opj(cwd_site, sdir)):
+                    dirs.remove(sdir)
+                else:
+                    os.mkdir(opj(dir_out, cwd_site, sdir))
         for f in files:
             if re.search(opts.ignore, opj(cwd_site, f)):
                 pass
@@ -569,15 +571,17 @@ def build(project, opts):
                         f_dst = opj(dir_out, cwd_site, f)
                         f_dst = '%s.%s' % (os.path.splitext(f_dst)[0], ext)
                         print('info   : convert %s (%s)' % (f_src, func.__name__))
-                        func(f_src, f_dst)
+                        if not opts.dry_run:
+                            func(f_src, f_dst)
                         break
                 else:
-                    src = opj(cwd, f)
-                    try:
-                        shutil.copy(src, opj(dir_out, cwd_site))
-                    except OSError:
-                        # some filesystems like FAT won't allow shutil.copy
-                        shutil.copyfile(src, opj(dir_out, cwd_site, f))
+                    if not opts.dry_run:
+                        src = opj(cwd, f)
+                        try:
+                            shutil.copy(src, opj(dir_out, cwd_site))
+                        except OSError:
+                            # some filesystems like FAT won't allow shutil.copy
+                            shutil.copyfile(src, opj(dir_out, cwd_site, f))
 
     pages.sort(key=lambda p: int(p.get("sval", "0")))
 
@@ -641,10 +645,14 @@ def build(project, opts):
         # write HTML page
         fname = page.fname.replace(dir_in, dir_out)
         fname = re.sub(MKD_PATT, ".html", fname)
-        with codecs.open(fname, 'w', opts.output_enc) as fp:
-            fp.write(out)
+        if not opts.dry_run:
+            with codecs.open(fname, 'w', opts.output_enc) as fp:
+                fp.write(out)
 
-    print("success: built project")
+    if opts.dry_run:
+        print("success: built project (dry-run)")
+    else:
+        print("success: built project")
 
 # =============================================================================
 # serve site
@@ -703,6 +711,8 @@ def options():
                   help="encoding of output pages (default: utf-8)")
     og.add_option("", "--filename-enc", default="utf-8", metavar="ENC",
                   help="encoding of file names (default: utf-8)")
+    op.add_option("" , "--dry-run", action="store_true", default=False,
+                  help="go through the rendering process without actually outputting/deleting any files")
     op.add_option_group(og)
 
     og = optparse.OptionGroup(op, "Serve options")
